backend/requirements.txt
pyshark==0.4.3.0
requests==2.31.0
PyYAML==6.0
pandas==2.2.2
scikit-learn==1.3.2
flask==2.4.2
flask-cors==4.0.0
sqlalchemy==2.0.20
psutil==5.9.5
gunicorn==21.2.0
